Support for timestamp change script numbering
Made modifications to allow use of timestamp numbering instead of sequential numbering (sequential numbering can still be used).  Modifications include:
* Converted change script number from int to long
* Created an ant task to automatically create a new timestamped change script
* Created a command line interface that can be run to create a new timestamped change script.

The timestamp is currently not configurable and in the format yyyyMMddHHmms.

Thanks to Jim Bogan for theses changes (with minor change to enable tests to run under intellij by me).

+++ dbdeploy-core/src/main/java/com/dbdeploy/database/changelog/DatabaseSchemaVersionManager.java
@@ -25,15 +25,15 @@ public class DatabaseSchemaVersionManager implements AppliedChangesProvider {
         this.changeLogTableName = changeLogTableName;
     }
 
-	public List<Integer> getAppliedChanges() {
+	public List<Long> getAppliedChanges() {
 		try {
 			ResultSet rs = queryExecuter.executeQuery(
 					"SELECT change_number FROM " + changeLogTableName + "  ORDER BY change_number");
 
-			List<Integer> changeNumbers = new ArrayList<Integer>();
+			List<Long> changeNumbers = new ArrayList<Long>();
 
 			while (rs.next()) {
-				changeNumbers.add(rs.getInt(1));
+				changeNumbers.add(rs.getLong(1));
 			}
 
 			rs.close();
